requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
cryptography
pyjwt
sqlalchemy
databases
aiosqlite